FormParserTraitを追加し、cfFormMailerクラスにフォーム解析機能を統合
